feat: Add user ID to created links
This commit modifies the `createLink` controller in the `link` directory to include the `userId` in the `Slug` model when creating a new link. This ensures that the link is associated with the correct user.

The `created_by` field in the `Slug` model is updated to store the user ID as an integer and has a foreign key reference to the `User` model.

This change allows for better tracking and management of links created by each user.